split demo view into separate react components
